Back/requirements.txt
discord.py
python-docx
PyMuPDF